requirements.txt
tensorflow>=2.10
opencv-python-headless>=4.5
pillow>=9.0
gradio>=3.39
numpy>=1.24
gunicorn
keras

